Create README.md
In this project, I used my python skills to code and decode messages. This file can be downloaded and tested on the Jupyter notebook.
This project was divided into 6 tasks. Steps 1 and 2 involved decoding a Caesar Cypher message using python, Steps 3 and  4 involved creating functions that can decode information with and without the offset number. Finally steps 5 and 6 involved decoding the Vigenère Cipher. Check it out